+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import './Reset.css';
 import './App.css';
-import Navbar from './componentes/Navbar';
+import Navbar from './components/Navbar/Navbar';
 import { Routes, Route } from 'react-router-dom';
 import Home from './pages/Home';
 import SinglePosts from './pages/SinglePosts';
